+++ usr.sbin/bhyvectl/Makefile
@@ -1,17 +1,18 @@
-#
-#
-
 .include <src.opts.mk>
+
+.PATH: ${.CURDIR}/${MACHINE_CPUARCH}
 
 PROG=	bhyvectl
 SRCS=	bhyvectl.c
 PACKAGE=	bhyve
 
+.include "${MACHINE_CPUARCH}/Makefile.inc"
+
 MAN=	bhyvectl.8
 
 LIBADD=	vmmapi util
 
-CFLAGS+= -I${SRCTOP}/sys/amd64/vmm
+CFLAGS+= -I${.CURDIR} -I${SRCTOP}/sys/amd64/vmm
 
 .if ${MK_BHYVE_SNAPSHOT} != "no"
 LIBADD+= nv
