+++ usr.sbin/bhyvectl/amd64/Makefile.inc
@@ -1,0 +1,1 @@
+SRCS+=	bhyvectl_machdep.c
